multi-sdf

+++ dsdf/netowrk.py
@@ -45,11 +45,11 @@ class DSDFNetwork(nn.Module):
     def forward(self, x):
         # x: [B, 4]
 
-        import ipdb; ipdb.set_trace()
+        #import ipdb; ipdb.set_trace()
         #x = x.double()
 
-        x = self.encoder(x)
-        #x = x.double()
+        x = self.encoder(x).float()
+        #x = x
 
         h = x
         for l in range(self.num_layers):
